Global exception handler (#22)
* Update package references and versions

- Added System.Text.Json to Askstatus.Web.API.Tests.csproj and Askstatus.Web.App.csproj
- Updated multiple package versions in Directory.Packages.props:
  - FluentAssertions to 7.0.0
  - FluentAssertions.Analyzers to 0.34.1
  - FluentValidation to 11.11.0
  - JunitXml.TestLogger to 4.1.0
  - Microsoft.AspNetCore.Authorization to 9.0.0
  - Microsoft.AspNetCore.Components.WebAssembly to 8.0.11
  - Microsoft.AspNetCore.Components.WebAssembly.Authentication to 8.0.11
  - Microsoft.AspNetCore.Components.WebAssembly.DevServer to 8.0.11
  - Microsoft.AspNetCore.Identity.EntityFrameworkCore to 8.0.11
  - Microsoft.AspNetCore.Mvc.Testing to 8.0.11
  - Microsoft.EntityFrameworkCore to 8.0.11
  - Microsoft.EntityFrameworkCore.Design to 8.0.11
  - Microsoft.EntityFrameworkCore.InMemory to 8.0.11
  - Microsoft.EntityFrameworkCore.Sqlite to 8.0.11
  - Microsoft.EntityFrameworkCore.Tools to 8.0.11
  - Microsoft.Extensions.Http to 8.0.1
  - Microsoft.NET.Test.Sdk to 17.12.0
  - MudBlazor to 7.15.0
  - Refit.Newtonsoft.Json to 8.0.0
  - Serilog.AspNetCore to 8.0.3
  - Swashbuckle.AspNetCore to 7.1.0
  - Added System.Text.Json 8.0.5
- Removed several outdated package versions

* Downgrade Microsoft.AspNetCore.Authorization package

The version of the Microsoft.AspNetCore.Authorization package has been downgraded from 9.0.0 to 8.0.11 in the Directory.Packages.props file. This change addresses compatibility issues or bugs that necessitate using an older version of the package.

* Update MudAppBar styles in MainLayout.razor

The MudAppBar component in the MainLayout.razor file has been updated to include specific styling attributes. The top MudAppBar now has `Style="margin-left: 0px; width: 100%"`, ensuring it spans the full width of the container and aligns to the left margin. Similarly, the bottom MudAppBar has been modified to include `Style="top: auto; bottom: 0; margin-left: 0px; width: 100%"`, ensuring it also spans the full width and aligns to the left margin.

* Add new API endpoints and global exception handling

Introduced new GET and POST endpoints for roles and permissions in the Askstatus.Web.API project. Added a GlobalExceptionHandler class to handle exceptions globally, logging them and returning a ProblemDetails response with a 500 status code. Registered the GlobalExceptionHandler and ProblemDetails services in the dependency injection container and configured the application to use the global exception handler middleware.